Forcing Kiwi to play until >= 17
Kiwi will only stop playing Blackjack, when both potential scores are >= 17. This is to avoid a case where Kiwi has a 15 and 22, and the player has <=15. Originally this case was added to stop players from not hitting and staying on low numbers. Now Kiwi will "stand" when both potential scores are >= 17, and play with the score that falls within 17 <= max(x,y) <= 21 range, where x and y are <= 21 and >= 17